implement error handling when creating resources for the first time. Should any erro occur, then delete all partially created resources if they exist.

+++ src/main/java/de/hsbo/kommonitor/datamanagement/api/impl/metadata/references/ReferenceManager.java
@@ -87,8 +87,10 @@ public class ReferenceManager {
 	}
 
 	public static void removeReferences(String indicatorId) {
-		indicatorRefRepo.deleteByIndicatorId(indicatorId);
-		georesourceRefRepo.deleteByMainIndicatorId(indicatorId);
+		if(indicatorRefRepo.existsByIndicatorId(indicatorId))
+			indicatorRefRepo.deleteByIndicatorId(indicatorId);
+		if(georesourceRefRepo.existsByMainIndicatorId(indicatorId))
+			georesourceRefRepo.deleteByMainIndicatorId(indicatorId);
 
 	}
 }
